LPS-97874 Rename kernel packages
These classes will now belong to the Layout app.
