+++ modules/apps/seo/seo-impl/src/main/java/com/liferay/seo/internal/SEOLinkManagerImpl.java
@@ -17,8 +17,8 @@ package com.liferay.seo.internal;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
 import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
-import com.liferay.portal.kernel.seo.SEOLink;
-import com.liferay.portal.kernel.seo.SEOLinkManager;
+import com.liferay.layouts.seo.kernel.SEOLink;
+import com.liferay.layouts.seo.kernel.SEOLinkManager;
 import com.liferay.portal.kernel.util.Html;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.seo.internal.configuration.SEOCompanyConfiguration;
