+++ modules/apps/seo/seo-test/src/testIntegration/java/com/liferay/seo/test/SEOLinkManagerTest.java
@@ -17,8 +17,8 @@ package com.liferay.seo.test;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.function.UnsafeRunnable;
 import com.liferay.portal.configuration.test.util.ConfigurationTemporarySwapper;
-import com.liferay.portal.kernel.seo.SEOLink;
-import com.liferay.portal.kernel.seo.SEOLinkManager;
+import com.liferay.layouts.seo.kernel.SEOLink;
+import com.liferay.layouts.seo.kernel.SEOLinkManager;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.HashMapDictionary;
