+++ portal-web/docroot/html/common/themes/init.jsp
@@ -18,8 +18,8 @@
 
 <%@ page import="com.liferay.portal.kernel.model.Portlet" %><%@
 page import="com.liferay.portal.kernel.model.portlet.PortletDependency" %><%@
-page import="com.liferay.portal.kernel.seo.SEOLink" %><%@
-page import="com.liferay.portal.kernel.seo.SEOLinkManagerUtil" %><%@
+page import="com.liferay.layouts.seo.kernel.SEOLink" %><%@
+page import="com.liferay.layouts.seo.kernel.SEOLinkManagerUtil" %><%@
 page import="com.liferay.portal.kernel.servlet.BrowserMetadata" %><%@
 page import="com.liferay.portal.kernel.upload.UploadServletRequestConfigurationHelperUtil" %><%@
 page import="com.liferay.portal.servlet.ComboServletStaticURLGenerator" %><%@
